@mvines 2021-01-17
Can you please explain a little more about what you've found with the threading here?  What does `.threads(1)` do?  Maybe links into the actual jsonrpc code would help.  Also do you know if the latest jsonrpc code behaves the same way?