@ryoqun 2021-01-17
some casual perf test:

small and fast reqs: no noticeable change.

BEFORE

```
$ ab -n 1000000 -c 1000 http://localhost:18899/health
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 100000 requests
Completed 200000 requests
Completed 300000 requests
Completed 400000 requests
Completed 500000 requests
Completed 600000 requests
Completed 700000 requests
Completed 800000 requests
Completed 900000 requests
Completed 1000000 requests
Finished 1000000 requests


Server Software:
Server Hostname:        localhost
Server Port:            18899

Document Path:          /health
Document Length:        2 bytes

Concurrency Level:      1000
Time taken for tests:   45.347 seconds
Complete requests:      1000000
Failed requests:        0
Total transferred:      77000000 bytes
HTML transferred:       2000000 bytes
Requests per second:    22052.30 [#/sec] (mean)
Time per request:       45.347 [ms] (mean)
Time per request:       0.045 [ms] (mean, across all concurrent requests)
Transfer rate:          1658.23 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0   23  63.8     19    1062
Processing:     0   22  11.9     21     237
Waiting:        0   15  11.6     15     230
Total:          0   45  65.0     42    1083

Percentage of the requests served within a certain time (ms)
  50%     42
  66%     46
  75%     47
  80%     48
  90%     51
  95%     53
  98%     55
  99%     60
 100%   1083 (longest request)
```

AFTER

```
$ ab -n 1000000 -c 1000 http://localhost:8899/health
This is ApacheBench, Version 2.3 <$Revision: 1843412 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 100000 requests
Completed 200000 requests
Completed 300000 requests
Completed 400000 requests
Completed 500000 requests
Completed 600000 requests
Completed 700000 requests
Completed 800000 requests
Completed 900000 requests
Completed 1000000 requests
Finished 1000000 requests


Server Software:
Server Hostname:        localhost
Server Port:            8899

Document Path:          /health
Document Length:        2 bytes

Concurrency Level:      1000
Time taken for tests:   42.470 seconds
Complete requests:      1000000
Failed requests:        0
Total transferred:      77000000 bytes
HTML transferred:       2000000 bytes
Requests per second:    23546.11 [#/sec] (mean)
Time per request:       42.470 [ms] (mean)
Time per request:       0.042 [ms] (mean, across all concurrent requests)
Transfer rate:          1770.56 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0   21  54.2     18    1050
Processing:     0   21  13.3     20     242
Waiting:        0   15  13.1     14     235
Total:          0   42  56.4     40    1246

Percentage of the requests served within a certain time (ms)
  50%     40
  66%     43
  75%     44
  80%     45
  90%     46
  95%     48
  98%     51
  99%     58
 100%   1246 (longest request)

```