@ryoqun 2021-01-17
heavy requests:

before:

```
$ echo '{"jsonrpc":"2.0", "id":1, "method":"getConfirmedBlock", "params":[61034542, "jsonParsed"]}' > post-data-get-confirmed-block  && ab -n 2000 -c 1000 -T "application/json" -p post-data-get-confirmed-block http://localhost:18899/
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 200 requests
Completed 400 requests
Completed 600 requests
Completed 800 requests
Completed 1000 requests
Completed 1200 requests
Completed 1400 requests
Completed 1600 requests
Completed 1800 requests
Completed 2000 requests
Finished 2000 requests


Server Software:        
Server Hostname:        localhost
Server Port:            18899

Document Path:          /
Document Length:        654837 bytes

Concurrency Level:      1000
Time taken for tests:   6.125 seconds
Complete requests:      2000
Failed requests:        0
Total transferred:      1309928000 bytes
Total body sent:        454000
HTML transferred:       1309674000 bytes
Requests per second:    326.52 [#/sec] (mean)
Time per request:       3062.586 [ms] (mean)
Time per request:       3.063 [ms] (mean, across all concurrent requests)
Transfer rate:          208847.42 [Kbytes/sec] received
                        72.38 kb/s sent
                        208919.80 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    8   8.0     11      20
Processing:    64 2214 983.3   2441    3855
Waiting:       64 2214 983.3   2438    3854
Total:         83 2222 976.9   2441    3855

Percentage of the requests served within a certain time (ms)
  50%   2441
  66%   2838
  75%   3016
  80%   3114
  90%   3363
  95%   3488
  98%   3600
  99%   3669
 100%   3855 (longest request)

```

after

```
$ echo '{"jsonrpc":"2.0", "id":1, "method":"getConfirmedBlock", "params":[61034542, "jsonParsed"]}' > post-data-get-confirmed-block  && ab -n 2000 -c 1000 -T "application/json" -p post-data-get-confirmed-block http://localhost:8899/
This is ApacheBench, Version 2.3 <$Revision: 1843412 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 200 requests
Completed 400 requests
Completed 600 requests
Completed 800 requests
Completed 1000 requests
Completed 1200 requests
Completed 1400 requests
Completed 1600 requests
Completed 1800 requests
Completed 2000 requests
Finished 2000 requests


Server Software:        
Server Hostname:        localhost
Server Port:            8899

Document Path:          /
Document Length:        654837 bytes

Concurrency Level:      1000
Time taken for tests:   5.341 seconds
Complete requests:      2000
Failed requests:        0
Total transferred:      1309928000 bytes
Total body sent:        452000
HTML transferred:       1309674000 bytes
Requests per second:    374.44 [#/sec] (mean)
Time per request:       2670.689 [ms] (mean)
Time per request:       2.671 [ms] (mean, across all concurrent requests)
Transfer rate:          239493.73 [Kbytes/sec] received
                        82.64 kb/s sent
                        239576.37 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    7   7.6      0      19
Processing:    50 1963 952.8   2075    4051
Waiting:       26 1963 952.7   2072    4050
Total:         55 1971 947.5   2080    4051

Percentage of the requests served within a certain time (ms)
  50%   2080
  66%   2492
  75%   2670
  80%   2837
  90%   3181
  95%   3370
  98%   3639
  99%   3928
 100%   4051 (longest request)

```