@ryoqun 2021-01-17
Also, I confirmed the rpc server indeed stuck at processing newer requests without this pr when it has encountered super heavy request.